third_party/blink/web_tests/platform/generic/media/W3C/video/events/event_canplaythrough-expected.txt

setting src attribute on autoplay video should trigger canplaythrough event

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS true is true
PASS successfullyParsed is true

TEST COMPLETE


spec reference
